Completed container implementation - needs logging and testing
